% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_indicators.R
\name{compute_indicators}
\alias{compute_indicators}
\title{compute_indicators}
\usage{
compute_indicators(dfList)
}
\description{
This function takes a list of dataframes with partially formatted data (e.g. from ILOSTAT) and calculates the indicators needed for the generate_rank function.
}
\details{
compute_indicators function
}
\keyword{compute}
\keyword{indicators}
