% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_helper.R
\name{filter_helper}
\alias{filter_helper}
\title{filter_helper}
\usage{
filter_helper(df, bygender, lastyear)
}
\description{
This function is used in rank_generator to filter observations on gender (for male, female, or total) and choose the most recent observation for each country, provided that country is no older than the year indicated in the function parameter "last year".
}
\details{
filter_helper function
}
\examples{
index <- filter_helper(neet, bygender = TRUE, lastyear= 2009)
}
\keyword{filter}
\keyword{helper}
