% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_helper.R
\name{impute_helper}
\alias{impute_helper}
\title{impute_helper}
\usage{
impute_helper(index)
}
\description{
This helper function imputes missing values conditional on the country having enough observations to be considered for the index.
}
\details{
impute_helper function
}
\keyword{helper}
\keyword{impute}
